requirements.txt
ursina
perlin-noise
Pillow
numpy
selenium
openai
easygui